requirements.txt
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
deprecation==2.1.0
ecdsa==0.19.0
exceptiongroup==1.2.2
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Login==0.6.3
gotrue==1.3.1
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
passlib==1.7.4
postgrest==0.11.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-discogs-client==2.7
python-dotenv==1.0.0
python-jose==3.3.0
realtime==1.0.6
requests==2.31.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
storage3==0.6.1
StrEnum==0.4.15
supabase==1.2.0
supafunc==0.2.3
typing_extensions==4.12.2
urllib3==2.3.0
websockets==12.0
Werkzeug==3.1.3
